NEU-Surface-Defect-Classification-main/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
jinja2==3.1.2
torch
torchvision
pillow

